# internal/server/server.go
package server

import (
	"context"
	"errors"
	"file_share/internal/deps"
	"file_share/internal/entity"
	"file_share/internal/handler/auth"
	"file_share/internal/handler/file"
	"file_share/internal/handler/folder"
	middleware "file_share/internal/handler/middlewares/auth"
	"file_share/internal/handler/scan"
	videosHandler "file_share/internal/handler/videos"
	"net/http"

	"github.com/gin-gonic/gin"
)

type Server struct {
	videosHandler  *videosHandler.Handler
	folderHandler  *folder.Handler
	authHandler    *auth.Handler
	scanHandler    *scan.Handler
	fileHandler    *file.Handler
	logger         deps.Logger
	addr           string
	server         *http.Server
	authMiddleware *middleware.Middleware
}

func NewServer(
	videosHandler *videosHandler.Handler,
	folderHandler *folder.Handler,
	scanHandler *scan.Handler,
	logger deps.Logger,
	addr string,
	authHandler *auth.Handler,
	authMiddleware *middleware.Middleware,
	fileHandler *file.Handler,
) *Server {
	s := &Server{
		videosHandler:  videosHandler,
		folderHandler:  folderHandler,
		scanHandler:    scanHandler,
		fileHandler:    fileHandler,
		logger:         logger,
		addr:           addr,
		authHandler:    authHandler,
		authMiddleware: authMiddleware,
	}

	router := s.Register(gin.Default())

	s.server = &http.Server{
		Addr:    s.addr,
		Handler: router,
	}

	return s
}

func (s *Server) Run(ctx context.Context) {
	s.logger.Info(ctx, "server starting", "address", s.addr)

	if err := s.server.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
		s.logger.Error(ctx, err)
	}
}

func (s *Server) Shutdown(ctx context.Context) error {
	return s.server.Shutdown(ctx)
}

func LiberalCORS(c *gin.Context) {
	c.Header("Access-Control-Allow-Origin", "*")
	c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")

	if c.Request.Method == "OPTIONS" {
		if len(c.Request.Header["Access-Control-Request-Headers"]) > 0 {
			c.Header("Access-Control-Allow-Headers",
				c.Request.Header["Access-Control-Request-Headers"][0])
		}
		c.AbortWithStatus(http.StatusOK)
	}
}

func (s *Server) Register(router *gin.Engine) *gin.Engine {
	router.Use(LiberalCORS)
	//router.Use(s.authMiddleware.Apply)

	api := router.Group("/api")
	{
		api.GET("/health")

		authGroup := api.Group("/auth")
		{
			authGroup.POST("/login", s.authHandler.Login)
			authGroup.POST("/register", s.authHandler.Register)
			authGroup.POST("/me", s.authMiddleware.Apply(entity.RoleUser, entity.RoleAdmin), s.authHandler.Me)
		}

		videos := api.Group("/videos")
		{
			videos.GET("/", s.videosHandler.GetAll)
			videos.GET("/:videoId", s.videosHandler.GetVideo)
			videos.GET("/:videoId/stream", s.videosHandler.Stream)
			videos.HEAD("/:videoId/stream", s.videosHandler.Stream)
			videos.GET("/:videoId/poster", s.videosHandler.GetPoster)
			videos.PUT("/:videoId/poster", s.authMiddleware.Apply(entity.RoleAdmin), s.videosHandler.UpdatePoster)

		}

		files := api.Group("/files")
		{
			files.GET("/:id/download", s.fileHandler.Download)
			files.HEAD("/:id/download", s.fileHandler.Download)
		}

		folders := api.Group("/folders")
		{
			folders.GET("/", s.folderHandler.GetAll)
			folders.POST("/", s.authMiddleware.Apply(entity.RoleAdmin), s.folderHandler.CreateRootFolder)
			folders.GET("/:folderId", s.folderHandler.GetFolder)
			folders.GET("/root/entries", s.authMiddleware.Apply(entity.RoleUser, entity.RoleAdmin), s.folderHandler.GetRootFolderEntries)
			folders.PATCH("/:folderId", s.authMiddleware.Apply(entity.RoleAdmin), s.folderHandler.UpdateFolder)
			folders.DELETE("/:folderId", s.authMiddleware.Apply(entity.RoleAdmin), s.folderHandler.DeleteFolder)
			folders.GET("/:folderId/entries", s.folderHandler.GetFoldersEntries)
			folders.GET("/:folderId/rescan", s.authMiddleware.Apply(entity.RoleAdmin), s.folderHandler.FolderRescan)
		}

		api.GET("/scan-jobs/:jobId", s.scanHandler.GetScanJob)
	}

	return router
}